preview.js→投稿とプロフィールに切り分けて、特定のファイルにのみ読み込ませるように
