+++ app/javascript/packs/application.js
@@ -8,5 +8,4 @@ require('@rails/ujs').start();
 require('@rails/activestorage').start();
 require('channels');
 require('./like');
-require('./preview');
 require('./tab');
