+++ app/views/posts/_form.html.haml
@@ -17,3 +17,6 @@
         = image_tag post.image, class: 'form__new-image'
       %button.form__button.button-primary
         = f.submit "投稿"
+
+
+= javascript_pack_tag 'preview-image'
