+++ app/views/profiles/edit.html.haml
@@ -19,3 +19,6 @@
         = image_tag @profile.avatar, class: 'form__new-image'
       %button.form__button.button-primary
         = f.submit "更新する"
+
+
+= javascript_pack_tag 'preview-avatar'
